modefied booking card for discount display and change added the final code for history booking page.
