+++ booking/src/_components/BookingHistory/BookingHistoryClient.tsx
@@ -1,4 +1,3 @@
-// src/StoryComponents/BookingHistory/BookingHistoryClient.tsx
 'use client';
 
 import React, { useState, useEffect, useRef, useCallback } from 'react';
